# src/main/resources/swagger/openapi.yaml
openapi: 3.0.0
info:
  title: Dev News
  version: "0.1"
  description: Dev News - API
...
servers:
  - url: /api/v1
paths:
...
  /users:
    get:
      tags:
        - User
      operationId: getUsers
      security:
        - userAuth: []
      parameters:
        - in: query
          name: name
          required: false
          schema:
            type: string
        - in: query
          name: username
          required: false
          schema:
            type: string
        - in: query
          name: email
          required: false
          schema:
            type: string
        - $ref: "#/components/parameters/PageQueryParameter"
        - $ref: "#/components/parameters/SizeQueryParameter"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserModel"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  "/users/{userId}":
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - User
      operationId: getUser
      security:
        - userAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserModel"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags:
        - User
      operationId: updateUser
      security:
        - userAuth: []
      requestBody:
        $ref: "#/components/requestBodies/UserRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserModel"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

    delete:
      tags:
        - User
      operationId: deleteUser
      security:
        - userAuth: []
      responses:
        "204":
          description: No Content
        "401":
          $ref: "#/components/responses/Unauthorized"
  /users/{userId}/image:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags:
        - User
      operationId: uploadImage
      security:
        - userAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadModel"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"

  # ========================================
  # ===== Category API
  # ========================================
...
components:
...
  parameters:
    PageQueryParameter:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    SizeQueryParameter:
      name: size
      in: query
      schema:
        type: integer
        minimum: 10
        maximum: 30
        default: 10

  # ========================================
  # ===== RequestBodies
  # ========================================
  requestBodies:
...
    UserRequest:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/UserModel"
      required: true
...
  responses:
    BadRequest:
      description: Bad Request
    Unauthorized:
      description: Unauthorized
...
    NotFound:
      description: Not Found
...
    UnprocessableEntity:
      description: Unprocessable Entity
...
  securitySchemes:
    userAuth:
      type: apiKey
      in: header
      name: Authorization
  schemas:
...
    UserRole:
      type: string
      enum:
        - USER
        - ADMIN
        - MODERATOR
        - PUBLISHER
...
    UserModel:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
          format: string
        email:
          type: string
          format: email
        phone:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        imageUrl:
          type: string
        bgUrl:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        birthday:
          type: string
        country:
          type: string
        city:
          type: string
        createdAt:
          type: string
...
    UploadModel:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
        user_id:
          type: string
          format: uuid
        post_id:
          type: string
          format: uuid